refactor : int -> BigDecimal로 변경
2022. 02. 22.
요금 필드 BigDecimal로 변경

Resolves : loop-study
See also : junwoochoi

+++ src/main/java/nextstep/subway/applicaion/dto/LineRequest.java
@@ -1,5 +1,7 @@
 package nextstep.subway.applicaion.dto;
 
+import java.math.BigDecimal;
+
 public class LineRequest {
     private String name;
     private String color;
@@ -7,7 +9,7 @@ public class LineRequest {
     private Long downStationId;
     private int distance;
     private int duration;
-    private int additionalFare;
+    private BigDecimal additionalFare;
 
     public String getName() {
         return name;
@@ -33,7 +35,7 @@ public class LineRequest {
         return duration;
     }
 
-    public int getAdditionalFare() {
+    public BigDecimal getAdditionalFare() {
         return additionalFare;
     }
 }
