+++ src/main/java/nextstep/subway/applicaion/dto/PathResponse.java
@@ -2,6 +2,7 @@ package nextstep.subway.applicaion.dto;
 
 import nextstep.subway.domain.Path;
 
+import java.math.BigDecimal;
 import java.util.List;
 import java.util.stream.Collectors;
 
@@ -9,9 +10,9 @@ public class PathResponse {
     private List<StationResponse> stations;
     private int distance;
     private int duration;
-    private int fare;
+    private BigDecimal fare;
 
-    public PathResponse(List<StationResponse> stations, int distance, int duration, int fare) {
+    public PathResponse(List<StationResponse> stations, int distance, int duration, BigDecimal fare) {
         this.stations = stations;
         this.distance = distance;
         this.duration = duration;
@@ -28,7 +29,7 @@ public class PathResponse {
                 .collect(Collectors.toList());
         int distance = path.extractDistance();
         int duration = path.extractDuration();
-        int fare = shortestPath.extractFare();
+        BigDecimal fare = shortestPath.extractFare();
 
         return new PathResponse(stations, distance, duration, fare);
     }
@@ -45,7 +46,7 @@ public class PathResponse {
         return duration;
     }
 
-    public int getFare() {
+    public BigDecimal getFare() {
         return fare;
     }
 }
